Added discrepancies between cvsimport and old cvswork
CVS is not the most precise tool.  These changes appear to have not made it into the CVS
repository.  This is either due to CVS itself, or NetBeans incorrectly reporting things
as up-to-date.

Either way, I have reviewed all these changes versus the current production system, and
they are correct.